python/requirements.txt
mysql-connector-python
openpyxl
fastapi
uvicorn